style-fix

+++ contracts/ESN/CertiDapp/BuildSurvey.sol
@@ -24,12 +24,12 @@ contract BuildSurvey is Governable, RegistryDependent {
         bool isPublic;
         //   uint[][10] feedback = new uint[][10](question);
     }
-    enum State { START, PRIVATE, COMPLETE ,PUBLISH}
+    enum State { START, PRIVATE, COMPLETE, PUBLISH }
 
     mapping(bytes32 => mapping(address => uint8)) public accessUser; // 0- no access , 1- access can vote,  2-  access already voted
     mapping(bytes32 => Survey) public surveys;
-    mapping(address => uint) public Incentives;
-    mapping(bytes32 => uint) public Funds;
+    mapping(address => uint256) public Incentives;
+    mapping(bytes32 => uint256) public Funds;
     // mapping(address => bool) public KYC; //proxy
     // kyc-1 needed for user .
 
@@ -56,9 +56,10 @@ contract BuildSurvey is Governable, RegistryDependent {
     //     KYC[user] = true;
     // }
 
-     function announceIncentive(uint _value) public{
+    function announceIncentive(uint256 _value) public {
         Incentives[msg.sender] = _value;
     }
+
     function addSurvey(
         string memory _title,
         string memory _surveyTitle,
@@ -88,7 +89,7 @@ contract BuildSurvey is Governable, RegistryDependent {
     }
 
     function addUsers(bytes32 _survey, address[] memory users) public payable onlyKycApproved {
-        require(msg.value == users.length,"Insufficient Funds");
+        require(msg.value == users.length, "Insufficient Funds");
         Funds[_survey] = Funds[_survey] + users.length;
         for (uint256 i = 0; i < users.length; i++) {
             accessUser[_survey][users[i]] = 1;
@@ -101,16 +102,17 @@ contract BuildSurvey is Governable, RegistryDependent {
         if (surveys[_survey].isPublic == false) {
             require(accessUser[_survey][msg.sender] != 1, "You have no access for this survey");
             Funds[_survey] = Funds[_survey] - 1;
-            msg.sender.transfer(1 ether );
+            msg.sender.transfer(1 ether);
         }
         require(accessUser[_survey][msg.sender] != 2, "You have already voted  for this survey");
         // surveys[_survey].feedback.push(_feedback);
         emit SentSurvey(_survey, _feedback);
-        require(msg.value == 1 ether,"Insufficient Funds");
+        require(msg.value == 1 ether, "Insufficient Funds");
         accessUser[_survey][msg.sender] = 2;
     }
-    function collectFunds(bytes32 _survey) public payable{
-        require(surveys[_survey].author == msg.sender,"You are not Authorized");
+
+    function collectFunds(bytes32 _survey) public payable {
+        require(surveys[_survey].author == msg.sender, "You are not Authorized");
         msg.sender.transfer(Funds[_survey]);
     }
 }
